requirements.txt
pyserial==2.7
mosquitto==1.2.3
bencode==1.0
docopt==0.6.2
